+++ CHANGELOG.md
@@ -7,6 +7,10 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased]
 
+### Removed
+
+- agentty: remove the `/qe:check` prompt slash command.
+
 ## [v0.12.7] - 2026-07-09
 
 ### Added